ID Patterns Edit
I implemented the new work flow and updated all the notation in the appinfo rmd as well as the figures.

Despite updating the the .R simulator function in the R file and in the simulatorfunction file and zip file the "Simulation Settings" still shows the wrong notation. 

Also, the dot notation would not work with the text printing function. I will look for a solution tomorrow.